+++ Swamp/Messages/PubSub/Subscriber/SubscribedSwampMessage.swift
@@ -11,10 +11,10 @@ import Foundation
 /// [SUBSCRIBED, requestId|number, subscription|number]
 class SubscribedSwampMessage: SwampMessage {
     
-    let requestId: Int
-    let subscription: Int
+    let requestId: UInt64
+    let subscription: UInt64
     
-    init(requestId: Int, subscription: Int) {
+    init(requestId: UInt64, subscription: UInt64) {
         self.requestId = requestId
         self.subscription = subscription
     }
@@ -22,8 +22,8 @@ class SubscribedSwampMessage: SwampMessage {
     // MARK: SwampMessage protocol
     
     required init(payload: [Any]) {
-        self.requestId = payload[0] as! Int
-        self.subscription = payload[1] as! Int
+        self.requestId = payload[0] as! UInt64
+        self.subscription = payload[1] as! UInt64
     }
     
     func marshal() -> [Any] {
